refactor: rename NotFoundError to Errors
