+++ src/domain/Errors.js
@@ -1,0 +1,7 @@
+class NotFoundError extends Error {
+}
+
+class UnableToCreatePassError extends Error {
+}
+
+export { NotFoundError, UnableToCreatePassError };
