--- src/domain/NotFoundError.js
@@ -1,4 +1,0 @@
-class NotFoundError extends Error {
-}
-
-export { NotFoundError };
