+++ src/infrastructure/ReservationRepository.js
@@ -1,5 +1,5 @@
 import { knex } from '../../db/knex-database-connection.js';
-import { NotFoundError } from '../domain/NotFoundError.js';
+import { NotFoundError } from '../domain/Errors.js';
 import { Reservation } from '../domain/Reservation.js';
 
 class ReservationRepository {
